Do not compile and add empty loaders
If the dependencies for a loader is not met we do not compile and add it to the loader registry.

Move all archive extraction code to archive loader.
